+++ src/scripts/transformation_data.py
@@ -1,4 +1,4 @@
-from src.data.transformation.data_transformation import DataTransformer
+from src.data.transformation.transformation_data import DataTransformer
 from pathlib import Path
 
 import argparse
